requirements.txt
asgiref==3.10.0
Django==4.2.26
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
